pyproject.toml
[tool.poetry]
name = "beauty-shop"
version = "0.1.0"
description = ""
authors = ["Vaipik <n.khylko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
django-debug-toolbar = "^4.1.0"
drf-spectacular = "^0.26.4"
psycopg = "^3.1.9"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
pillow = "^10.0.0"
django = "^4.2.4"
django-filter = "^23.2"
django-treebeard = {git = "https://github.com/Vaipik/django-treebeard.git"}
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"

pre-commit = "^3.3.3"
black = "^23.7.0"
ruff = "^0.0.280"
isort = "^5.12.0"
factory-boy = "^3.3.0"
pytest = "^7.4.2"
pytest-django = "^4.5.2"
pytest-factoryboy= "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
select = [
    "E",
    "F",
    "D"
]
exclude = ["settings.py", "admin.py", "apps.py", "migrations", "tests", "README.md"]
ignore = ["D100", "D104", "D105", "D106", "D107", "D104", "D401", "F821",]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
pythonpath = "src"
[tool.isort]
profile = "black"
